+++ components/Dialogs/src/private/MDCAlertControllerView+Private.h
@@ -26,6 +26,12 @@
 
 @property(nonatomic, nullable, weak) MDCAlertActionManager *actionManager;
 
+/** The scroll view that holds both the @c titleLabel and @c messageLabel. */
+@property(nonatomic, nonnull, strong) UIScrollView *contentScrollView;
+
+/** The scroll view that holds all of the buttons created for each action. */
+@property(nonatomic, nonnull, strong) UIScrollView *actionsScrollView;
+
 - (void)addActionButton:(nonnull MDCButton *)button;
 + (void)styleAsTextButton:(nonnull MDCButton *)button;
 
@@ -34,5 +40,17 @@
 - (CGSize)calculateActionsSizeThatFitsWidth:(CGFloat)boundingWidth;
 
 - (void)updateFonts;
+
+/**
+ Affects the fallback behavior for when a scaled font is not provided.
+
+ If @c YES, the font size will adjust even if a scaled font has not been provided for
+ a given @c UIFont property on this component.
+
+ If @c NO, the font size will only be adjusted if a scaled font has been provided.
+
+ Default value is @c YES.
+ */
+@property(nonatomic, assign) BOOL adjustsFontForContentSizeCategoryWhenScaledFontIsUnavailable;
 
 @end