+++ components/Tabs/src/MDCTabBar.m
@@ -16,7 +16,9 @@
 
 #import <MDFInternationalization/MDFInternationalization.h>
 
+#import "MDCTabBarExtendedAlignment.h"
 #import "MDCTabBarIndicatorTemplate.h"
+#import "MDCTabBarSizeClassDelegate.h"
 #import "MDCTabBarUnderlineIndicatorTemplate.h"
 #import "MaterialInk.h"
 #import "MaterialTypography.h"
@@ -53,24 +55,32 @@ static const CGFloat kBottomNavigationTitleImagePadding = 3;
 /// Height for the bottom divider.
 static const CGFloat kBottomNavigationBarDividerHeight = 1;
 
-static MDCItemBarAlignment MDCItemBarAlignmentForTabBarAlignment(MDCTabBarAlignment alignment) {
+static MDCItemBarAlignment MDCItemBarAlignmentForTabBarAlignment(
+    MDCTabBarExtendedAlignment alignment) {
   switch (alignment) {
-    case MDCTabBarAlignmentCenter:
+    case MDCTabBarExtendedAlignmentCenter:
       return MDCItemBarAlignmentCenter;
 
-    case MDCTabBarAlignmentLeading:
+    case MDCTabBarExtendedAlignmentLeading:
       return MDCItemBarAlignmentLeading;
 
-    case MDCTabBarAlignmentJustified:
+    case MDCTabBarExtendedAlignmentJustified:
       return MDCItemBarAlignmentJustified;
 
-    case MDCTabBarAlignmentCenterSelected:
+    case MDCTabBarExtendedAlignmentBestEffortJustified:
+      return MDCItemBarAlignmentBestEffortJustified;
+
+    case MDCTabBarExtendedAlignmentCenterSelected:
       return MDCItemBarAlignmentCenterSelected;
   }
 
   NSCAssert(0, @"Invalid alignment value %ld", (long)alignment);
   return MDCItemBarAlignmentLeading;
 }
+
+@interface MDCTabBar ()
+@property(nonatomic, weak, nullable) id<MDCTabBarSizeClassDelegate> sizeClassDelegate;
+@end
 
 @interface MDCTabBar () <MDCItemBarDelegate>
 @end
@@ -140,9 +150,11 @@ static MDCItemBarAlignment MDCItemBarAlignmentForTabBarAlignment(MDCTabBarAlignm
 
   // Create item bar.
   _itemBar = [[MDCItemBar alloc] initWithFrame:self.bounds];
+  _itemBar.tabBar = self;
   _itemBar.autoresizingMask = (UIViewAutoresizingFlexibleWidth | UIViewAutoresizingFlexibleHeight);
   _itemBar.delegate = self;
-  _itemBar.alignment = MDCItemBarAlignmentForTabBarAlignment(_alignment);
+  _itemBar.alignment =
+      MDCItemBarAlignmentForTabBarAlignment((MDCTabBarExtendedAlignment)_alignment);
   [self addSubview:_itemBar];
 
   CGFloat dividerTop = CGRectGetMaxY(self.bounds) - kBottomNavigationBarDividerHeight;
@@ -583,7 +595,9 @@ static MDCItemBarAlignment MDCItemBarAlignmentForTabBarAlignment(MDCTabBarAlignm
 - (void)internalSetAlignment:(MDCTabBarAlignment)alignment animated:(BOOL)animated {
   if (_alignment != alignment) {
     _alignment = alignment;
-    [_itemBar setAlignment:MDCItemBarAlignmentForTabBarAlignment(_alignment) animated:animated];
+    [_itemBar
+        setAlignment:MDCItemBarAlignmentForTabBarAlignment((MDCTabBarExtendedAlignment)_alignment)
+            animated:animated];
   }
 }