+++ components/Tabs/src/private/MDCItemBar.h
@@ -17,6 +17,7 @@
 #import "MDCItemBarAlignment.h"
 
 @class MDCItemBarStyle;
+@class MDCTabBar;
 @protocol MDCItemBarDelegate;
 
 /**
@@ -29,6 +30,14 @@
 
 /** Return the default height for the item bar given a style. */
 + (CGFloat)defaultHeightForStyle:(nonnull MDCItemBarStyle *)style;
+
+/**
+ Link back to this item bar's owning MDCTabBar so it can interrogate the tab bar's size class
+ delegate for a horizontal size class.
+
+ @note This property may be removed in the near future.
+ */
+@property(nonatomic, weak, nullable) MDCTabBar *tabBar;
 
 /**
  Items displayed in the item bar.