+++ components/schemes/Typography/src/MDCTypographyScheme.m
@@ -26,6 +26,8 @@
 - (instancetype)initWithDefaults:(MDCTypographySchemeDefaults)defaults {
   self = [super init];
   if (self) {
+    _useCurrentContentSizeCategoryWhenApplied = NO;
+
     switch (defaults) {
       case MDCTypographySchemeDefaultsMaterial201804:
 #if defined(__IPHONE_8_2)
@@ -62,7 +64,6 @@
         _button = [UIFont systemFontOfSize:14.0];
         _overline = [UIFont systemFontOfSize:12.0];
 #endif
-        _mdc_adjustsFontForContentSizeCategory = NO;
         break;
       case MDCTypographySchemeDefaultsMaterial201902:
 #if defined(__IPHONE_8_2)
@@ -99,8 +100,6 @@
         _button = [UIFont systemFontOfSize:14.0];
         _overline = [UIFont systemFontOfSize:12.0];
 #endif
-
-        _mdc_adjustsFontForContentSizeCategory = YES;
 
         // Attach a sizing curve to all fonts
         MDCFontScaler *fontScaler =
@@ -182,6 +181,14 @@
   copy.mdc_adjustsFontForContentSizeCategory = self.mdc_adjustsFontForContentSizeCategory;
 
   return copy;
+}
+
+- (BOOL)mdc_adjustsFontForContentSizeCategory {
+  return self.useCurrentContentSizeCategoryWhenApplied;
+}
+
+- (void)setMdc_adjustsFontForContentSizeCategory:(BOOL)mdc_adjustsFontForContentSizeCategory {
+  self.useCurrentContentSizeCategoryWhenApplied = mdc_adjustsFontForContentSizeCategory;
 }
 
 @end